requirements/base.txt
Django==3.2.5
djangorestframework==3.12.4


# third party dependencies
psycopg2==2.9.1
django-extensions==3.1.3
django-mptt==0.12.0
django-thumbnails==0.6.0
redis==3.5.3
django-extensions==3.1.3
Pillow==8.3.0
django-countries==7.2.1
requests==2.26.0
